# Optimize `TypeMemberReflector`