@jnyrup 2023-09-25
The file still has 100% line and branch coverage.
The reduction is because we now do the same in less code.

Duplicate check of `!field.IsPrivate`
![image](https://github.com/fluentassertions/fluentassertions/assets/919634/27171a07-732a-40e5-9f26-10813a7f0523)

We null-Guarded `GetMethod` in `IsExplicitImplementation` and later removing members without a getter.
![image](https://github.com/fluentassertions/fluentassertions/assets/919634/551d11ce-e309-4ad9-9c4e-f13f6237c846)