@jnyrup 2023-09-27
> What about the Qodana issue?

Hadn't noticed that.

Qodana cannot see that `HasGetter` filters out properties without a `GetMethod`.

So either the code must be restructured to make that visible to Qodana, or we add `!` in `IsExplicitImplementation`