@dennisdoomen 2023-09-28
> So either the code must be restructured to make that visible to Qodana, or we add `!` in `IsExplicitImplementation`

That's what my pragmatic mind would do ;-)